experiments/unet/requirements.txt
absl-py==0.13.0
aiohttp==3.8.1
aiosignal==1.3.1
albumentations==1.0.0
anyio==3.3.1
argon2-cffi-bindings==21.2.0
argon2-cffi==21.1.0
astor==0.8.1
asttokens==2.1.0
async-timeout==4.0.2
attrs==21.2.0
babel==2.9.1
backcall==0.2.0
bleach==4.1.0
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography==3.4.7
cycler==0.10.0
cython==0.29.32
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
fire==0.4.0
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
gitdb==4.0.7
gitpython==3.1.27
google-auth-oauthlib==0.4.6
google-auth==1.35.0
grpcio==1.40.0
huggingface-hub==0.11.0
idna==2.10
imageio==2.21.1
importlib-metadata==5.0.0
ipykernel==6.4.1
ipython-genutils==0.2.0
ipython==7.27.0
ipywidgets==7.6.5
jedi==0.18.0
jinja2==3.0.1
joblib==1.2.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-console==6.4.4
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyter==1.0.0
jupyterlab-language-pack-zh-cn==0.0.2.dev0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.2
jupyterlab==3.1.12
kiwisolver==1.3.2
kornia==0.5.8
lightning-utilities==0.3.0
markdown==3.3.4
markupsafe==2.0.1
matplotlib-inline==0.1.3
matplotlib==3.4.3
mistune==0.8.4
monai==0.5.2
msgpack==1.0.4
multidict==6.0.2
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8.6
notebook-shim==0.2.2
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
opencv-python-headless==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.0
paddle-bfloat==0.1.7
paddlepaddle-gpu==2.4.0
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.3.2
pip==22.2.2
platformdirs==2.5.4
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pydeprecate==0.3.0
pygments==2.10.0
pyopenssl==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
pysocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.3.2
pytz==2021.1
pywavelets==1.3.0
pyyaml==5.4.1
pyzmq==22.3.0
qtconsole==5.4.0
qtpy==2.3.0
qudida==0.0.4
regex==2022.10.31
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
requests==2.25.1
rsa==4.7.2
ruamel.yaml.clib==0.2.7
ruamel.yaml==0.17.21
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.1
send2trash==1.8.0
sentry-sdk==1.9.5
setproctitle==1.3.2
setuptools==51.1.0.post20201221
shortuuid==1.0.9
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.1
supervisor==4.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboard==2.6.0
tensorboardx==2.5.1
termcolor==2.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.1.1
tokenizers==0.13.2
torchmetrics==0.5.0
torchstat==0.0.7
torchsummary==1.5.1
tornado==6.1
tqdm==4.61.2
traitlets==5.1.0
traits==6.3.2
transformers==4.24.0
typing-extensions==3.10.0.2
tzdata==2022.2
urllib3==1.26.12
wandb==0.13.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
werkzeug==2.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
wincertstore==0.2
xmltodict==0.12.0
yarl==1.8.1
zipp==3.8.1